+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 [Full changelog](https://github.com/mozilla/glean/compare/v69.0.0...main)
 
+* General
+  * Accept 0-duration samples in timing_distribution metrics ([bug 2049040](https://bugzilla.mozilla.org/show_bug.cgi?id=2049040))
+
 # v69.0.0 (2026-06-22)
 
 [Full changelog](https://github.com/mozilla/glean/compare/v68.0.0...v69.0.0)